refactor(mcp): split tool dispatcher modules
Split the Anki MCP tool implementation into schema, response, validation, deck, model, and note modules while preserving dispatcher behavior.
